Modify ffmpeg cmd to eliminate re-encoding
Copies audio and video streams from h264/aac video source without re-encoding. This is a lazy hack to address redundant re-encoding, but breaks the ability to upload video formats that would otherwise need re-encoding.

In future, this should be written as a conditional with a fallback to original cmd for non h264/aac video.