Remove redundant creator name in dates of creation area
Where there are multiple creators, this displays the date redundantly with each creator or whichever creator has the date attached to it. Creators are already listed in archival description area, so this redundant.

This does not allow for other types of actors with non-creator roles to be shown in the main record panel, but will still show in 'related people and organizations' on right side.

Section commented out should desire to address later to include non-creator actors in dates area.